+++ java/integTest/src/test/java/glide/SharedCommandTests.java
@@ -50,8 +50,6 @@ import glide.api.models.commands.RangeOptions.RangeByLex;
 import glide.api.models.commands.RangeOptions.RangeByScore;
 import glide.api.models.commands.RangeOptions.ScoreBoundary;
 import glide.api.models.commands.RestoreOptions;
-import glide.api.models.commands.ScriptOptions;
-import glide.api.models.commands.ScriptOptionsGlideString;
 import glide.api.models.commands.SetOptions;
 import glide.api.models.commands.SortBaseOptions;
 import glide.api.models.commands.SortOptions;
@@ -992,11 +990,11 @@ public class SharedCommandTests {
     @SneakyThrows
     @ParameterizedTest(autoCloseArguments = false)
     @MethodSource("getClients")
-    public void non_UTF8_GlideString_test(BaseClient client) {
+    public void non_UTF8_GlideString_map(BaseClient client) {
         byte[] nonUTF8Bytes = new byte[] {(byte) 0xEE};
         GlideString key = gs(nonUTF8Bytes);
         GlideString hashKey = gs(UUID.randomUUID().toString());
-        GlideString hashNonUTF8Key = gs(new byte[] {(byte) 0xFF});
+        GlideString hashNonUTF8Key = gs(new byte[] {(byte) 0xDD});
         GlideString value = gs(nonUTF8Bytes);
         String stringField = "field";
         Map<GlideString, GlideString> fieldValueMap = Map.of(gs(stringField), value);
@@ -1021,6 +1019,150 @@ public class SharedCommandTests {
         assertEquals(
                 "Value not convertible to string: byte[] 13",
                 client.hget(hashNonUTF8Key, gs(stringField)).get().toString());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("getClients")
+    public void non_UTF8_GlideString_map_with_double(BaseClient client) {
+        byte[] nonUTF8Bytes = new byte[] {(byte) 0xEE};
+        GlideString key = gs(UUID.randomUUID().toString());
+        GlideString nonUTF8Key = gs(new byte[] {(byte) 0xEF});
+        Map<GlideString, Double> membersScores =
+                Map.of(gs(nonUTF8Bytes), 1.0, gs("two"), 2.0, gs("three"), 3.0);
+
+        // Testing map values using byte[] that cannot be converted to UTF-8 Strings.
+        assertEquals(3, client.zadd(key, membersScores).get());
+        assertThrows(
+                ExecutionException.class,
+                () -> client.zrange(key.toString(), new RangeByIndex(0, 1)).get());
+
+        // Testing keys for a map using byte[] that cannot be converted to UTF-8 Strings returns bytes.
+        assertEquals(3, client.zadd(nonUTF8Key, membersScores).get());
+        // No error is thrown as GlideString will be returned when arguments are GlideStrings.
+        assertDeepEquals(
+                new GlideString[] {gs(nonUTF8Bytes), gs("two"), gs("three")},
+                client.zrange(nonUTF8Key, new RangeByIndex(0, -1)).get());
+
+        // Converting non UTF-8 bytes result to String returns a message.
+        assertEquals(
+                "Value not convertible to string: byte[] 13",
+                client.zrange(nonUTF8Key, new RangeByIndex(0, -1)).get()[0].toString());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("getClients")
+    public void non_UTF8_GlideString_nested_array(BaseClient client) {
+        byte[] nonUTF8Bytes = new byte[] {(byte) 0xEE};
+        GlideString key = gs(UUID.randomUUID().toString());
+        GlideString nonUTF8Key = gs(new byte[] {(byte) 0xFF});
+        GlideString field = gs(nonUTF8Bytes);
+        GlideString value1 = gs(nonUTF8Bytes);
+        GlideString value2 = gs("foobar");
+        GlideString[][] entry = new GlideString[][] {{field, value1}, {field, value2}};
+
+        // Testing stream values using byte[] that cannot be converted to UTF-8 Strings.
+        client.xadd(key, entry).get();
+        assertThrows(
+                ExecutionException.class,
+                () -> client.xrange(key.toString(), InfRangeBound.MIN, InfRangeBound.MAX).get());
+
+        // Testing keys for a stream using byte[] that cannot be converted to UTF-8 Strings returns
+        // bytes.
+        GlideString streamId = client.xadd(nonUTF8Key, entry).get();
+        // No error is thrown as GlideString will be returned when arguments are GlideStrings.
+        Map<GlideString, GlideString[][]> expected =
+                Map.of(streamId, new GlideString[][] {{field, value1}, {field, value2}});
+        assertDeepEquals(
+                expected, client.xrange(nonUTF8Key, InfRangeBound.MIN, InfRangeBound.MAX).get());
+
+        // Converting non UTF-8 bytes result to String returns a message.
+        assertEquals(
+                "Value not convertible to string: byte[] 13",
+                client
+                        .xrange(nonUTF8Key, InfRangeBound.MIN, InfRangeBound.MAX)
+                        .get()
+                        .get(streamId)[0][0]
+                        .toString());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("getClients")
+    public void non_UTF8_GlideString_map_with_geospatial(BaseClient client) {
+        byte[] nonUTF8Bytes = new byte[] {(byte) 0xEE};
+        GlideString key = gs(UUID.randomUUID().toString());
+        GlideString nonUTF8Key = gs(new byte[] {(byte) 0xDF});
+        Map<GlideString, GeospatialData> membersToCoordinates = new HashMap<>();
+        membersToCoordinates.put(gs(nonUTF8Bytes), new GeospatialData(13.361389, 38.115556));
+        membersToCoordinates.put(gs("Catania"), new GeospatialData(15.087269, 37.502669));
+
+        // Testing geospatial values using byte[] that cannot be converted to UTF-8 Strings.
+        assertEquals(2, client.geoadd(key, membersToCoordinates).get());
+        assertThrows(
+                ExecutionException.class,
+                () ->
+                        client
+                                .geosearch(
+                                        key.toString(),
+                                        new CoordOrigin(new GeospatialData(15, 37)),
+                                        new GeoSearchShape(400, 400, GeoUnit.KILOMETERS))
+                                .get());
+
+        // Testing keys for geospatial using byte[] that cannot be converted to UTF-8 Strings returns
+        // bytes.
+        assertEquals(2, client.geoadd(nonUTF8Key, membersToCoordinates).get());
+        // No error is thrown as GlideString will be returned when arguments are GlideStrings.
+        assertTrue(
+                Set.of(new GlideString[] {gs(nonUTF8Bytes), gs("Catania")})
+                        .containsAll(
+                                Set.of(
+                                        client
+                                                .geosearch(
+                                                        nonUTF8Key,
+                                                        new CoordOrigin(new GeospatialData(15, 37)),
+                                                        new GeoSearchShape(400, 400, GeoUnit.KILOMETERS))
+                                                .get())));
+
+        // Converting non UTF-8 bytes result to String returns a message.
+        assertEquals(
+                "Value not convertible to string: byte[] 13",
+                client
+                        .geosearch(
+                                nonUTF8Key,
+                                new CoordOrigin(new GeospatialData(15, 37)),
+                                new GeoSearchShape(400, 400, GeoUnit.KILOMETERS))
+                        .get()[0]
+                        .toString());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("getClients")
+    public void non_UTF8_GlideString_map_of_arrays(BaseClient client) {
+        byte[] nonUTF8Bytes = new byte[] {(byte) 0xEE};
+        GlideString key = gs(UUID.randomUUID().toString());
+        GlideString nonUTF8Key = gs(new byte[] {(byte) 0xFE});
+        GlideString[] lpushArgs = {gs(nonUTF8Bytes), gs("two")};
+
+        // Testing map of arrays using byte[] that cannot be converted to UTF-8 Strings.
+        assertEquals(2, client.lpush(key, lpushArgs).get());
+        // trying to take a string from a key, but the key stores a non-string-compatible value
+        // decoding failed and value lost
+        assertThrows(
+                ExecutionException.class,
+                () -> client.lmpop(new String[] {key.toString()}, ListDirection.RIGHT).get());
+
+        // Testing map of arrays using byte[] that cannot be converted to UTF-8 Strings returns bytes.
+        assertEquals(2, client.lpush(nonUTF8Key, lpushArgs).get());
+        // No error is thrown as GlideString will be returned when arguments are GlideStrings.
+        var popResult = client.lmpop(new GlideString[] {nonUTF8Key}, ListDirection.RIGHT).get();
+        assertDeepEquals(Map.of(nonUTF8Key, new GlideString[] {gs(nonUTF8Bytes)}), popResult);
+
+        // Converting non UTF-8 bytes result to String returns a message.
+        assertEquals(
+                "Value not convertible to string: byte[] 13", popResult.get(nonUTF8Key)[0].toString());
     }
 
     @SneakyThrows
@@ -3131,122 +3273,6 @@ public class SharedCommandTests {
         assertTrue(0L <= persistAmount && persistAmount <= 10L);
         assertTrue(client.persist(gs(key)).get());
         assertEquals(-1L, client.ttl(key).get());
-    }
-
-    @SneakyThrows
-    @ParameterizedTest(autoCloseArguments = false)
-    @MethodSource("getClients")
-    public void invokeScript_test(BaseClient client) {
-        String key1 = UUID.randomUUID().toString();
-        String key2 = UUID.randomUUID().toString();
-
-        try (Script script = new Script("return 'Hello'", false)) {
-            Object response = client.invokeScript(script).get();
-            assertEquals("Hello", response);
-        }
-
-        try (Script script = new Script("return redis.call('SET', KEYS[1], ARGV[1])", false)) {
-            Object setResponse1 =
-                    client
-                            .invokeScript(script, ScriptOptions.builder().key(key1).arg("value1").build())
-                            .get();
-            assertEquals(OK, setResponse1);
-
-            Object setResponse2 =
-                    client
-                            .invokeScript(script, ScriptOptions.builder().key(key2).arg("value2").build())
-                            .get();
-            assertEquals(OK, setResponse2);
-        }
-
-        try (Script script = new Script("return redis.call('GET', KEYS[1])", false)) {
-            Object getResponse1 =
-                    client.invokeScript(script, ScriptOptions.builder().key(key1).build()).get();
-            assertEquals("value1", getResponse1);
-
-            // Use GlideString in option but we still expect nonbinary output
-            Object getResponse2 =
-                    client
-                            .invokeScript(script, ScriptOptionsGlideString.builder().key(gs(key2)).build())
-                            .get();
-            assertEquals("value2", getResponse2);
-        }
-    }
-
-    @SneakyThrows
-    @ParameterizedTest(autoCloseArguments = false)
-    @MethodSource("getClients")
-    public void script_large_keys_and_or_args(BaseClient client) {
-        String str1 = "0".repeat(1 << 12); // 4k
-        String str2 = "0".repeat(1 << 12); // 4k
-
-        try (Script script = new Script("return KEYS[1]", false)) {
-            // 1 very big key
-            Object response =
-                    client.invokeScript(script, ScriptOptions.builder().key(str1 + str2).build()).get();
-            assertEquals(str1 + str2, response);
-        }
-
-        try (Script script = new Script("return KEYS[1]", false)) {
-            // 2 big keys
-            Object response =
-                    client.invokeScript(script, ScriptOptions.builder().key(str1).key(str2).build()).get();
-            assertEquals(str1, response);
-        }
-
-        try (Script script = new Script("return ARGV[1]", false)) {
-            // 1 very big arg
-            Object response =
-                    client.invokeScript(script, ScriptOptions.builder().arg(str1 + str2).build()).get();
-            assertEquals(str1 + str2, response);
-        }
-
-        try (Script script = new Script("return ARGV[1]", false)) {
-            // 1 big arg + 1 big key
-            Object response =
-                    client.invokeScript(script, ScriptOptions.builder().arg(str1).key(str2).build()).get();
-            assertEquals(str2, response);
-        }
-    }
-
-    @SneakyThrows
-    @ParameterizedTest(autoCloseArguments = false)
-    @MethodSource("getClients")
-    public void invokeScript_gs_test(BaseClient client) {
-        GlideString key1 = gs(UUID.randomUUID().toString());
-        GlideString key2 = gs(UUID.randomUUID().toString());
-
-        try (Script script = new Script(gs("return 'Hello'"), true)) {
-            Object response = client.invokeScript(script).get();
-            assertEquals(gs("Hello"), response);
-        }
-
-        try (Script script = new Script(gs("return redis.call('SET', KEYS[1], ARGV[1])"), true)) {
-            Object setResponse1 =
-                    client
-                            .invokeScript(
-                                    script, ScriptOptionsGlideString.builder().key(key1).arg(gs("value1")).build())
-                            .get();
-            assertEquals(OK, setResponse1);
-
-            Object setResponse2 =
-                    client
-                            .invokeScript(
-                                    script, ScriptOptionsGlideString.builder().key(key2).arg(gs("value2")).build())
-                            .get();
-            assertEquals(OK, setResponse2);
-        }
-
-        try (Script script = new Script(gs("return redis.call('GET', KEYS[1])"), true)) {
-            Object getResponse1 =
-                    client.invokeScript(script, ScriptOptionsGlideString.builder().key(key1).build()).get();
-            assertEquals(gs("value1"), getResponse1);
-
-            // Use String in option but we still expect binary output (GlideString)
-            Object getResponse2 =
-                    client.invokeScript(script, ScriptOptions.builder().key(key2.toString()).build()).get();
-            assertEquals(gs("value2"), getResponse2);
-        }
     }
 
     @SneakyThrows